+++ src/features/diary/components/content/DiaryEditor/DiaryEditor.tsx
@@ -10,7 +10,7 @@ import { LinkPlugin } from '@lexical/react/LexicalLinkPlugin';
 import { OnChangePlugin } from '@lexical/react/LexicalOnChangePlugin';
 import { EditorState } from 'lexical/LexicalEditorState';
 
-import ToolbarPlugin from './plugins/ToolbarPlugin.tsx';
+import ToolbarPlugin from './plugins/ToolbarPlugin/index.tsx';
 import * as style from '@/features/diary/components/content/DiaryEditor/DiaryEditor.css.ts';
 import AutoLinkPlugin from './plugins/AutoLinkPlugin.tsx';
 import InitalPlugin from '@/features/diary/components/content/DiaryEditor/plugins/InitialPlugin.tsx';
